Split domain snapshot APIs to libvirtApi/snapshot.js
